错误修复：相册默认cover不显示

+++ src/main/java/com/example/photomanager/service/impl/AlbumServiceImpl.java
@@ -79,7 +79,9 @@ public class AlbumServiceImpl extends ServiceImpl<AlbumMapper, Album> implements
         }
 
         //添加到回收站
-        photoService.deletePhotos(list);
+        if (!list.isEmpty()){
+            photoService.deletePhotos(list);
+        }
 
         return removeById(id);
     }
